category/games/2026/02/20260209-063052-vector-rampaging-gorilla-city/pyproject.toml
[project]
name = "vector-rampaging-gorilla-city"
version = "0.1.0"
description = "Destroy high-rise buildings while dodging defense helicopters in this classic monster rampage simulation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.scripts]
vector-rampaging-gorilla-city = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "*.pyc",
    "__pycache__",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
